[core] allow to scope bindings per connection
