+++ packages/core/src/node/messaging/messaging-backend-module.ts
@@ -15,13 +15,19 @@
  ********************************************************************************/
 
 import { ContainerModule } from 'inversify';
-import { bindContributionProvider, ConnectionHandler } from '../../common';
+import { bindContributionProvider } from '../../common';
 import { BackendApplicationContribution } from '../backend-application';
-import { MessagingContribution } from './messaging-contribution';
+import { MessagingContribution, ConnectionContainerModule, MessagingContainer } from './messaging-contribution';
 import { MessagingService } from './messaging-service';
 
 export const messagingBackendModule = new ContainerModule(bind => {
-    bind(BackendApplicationContribution).to(MessagingContribution).inSingletonScope();
-    bindContributionProvider(bind, ConnectionHandler);
+    bindContributionProvider(bind, ConnectionContainerModule);
     bindContributionProvider(bind, MessagingService.Contribution);
+    bind(MessagingContribution).toDynamicValue(({ container }) => {
+        const child = container.createChild();
+        child.bind(MessagingContainer).toConstantValue(container);
+        child.bind(MessagingContribution).toSelf();
+        return child.get(MessagingContribution);
+    }).inSingletonScope();
+    bind(BackendApplicationContribution).toService(MessagingContribution);
 });
