+++ packages/debug/src/node/debug-backend-module.ts
@@ -17,11 +17,6 @@
 import { ConnectionHandler, JsonRpcConnectionHandler, bindContributionProvider, ILogger } from '@theia/core/lib/common';
 import { ContainerModule } from 'inversify';
 import {
-    DebugServiceImpl,
-    DebugAdapterSessionManager,
-    DebugAdapterContributionRegistry
-} from './debug-service-impl';
-import {
     DebugPath,
     DebugService
 } from '../common/debug-service';
@@ -30,27 +25,35 @@ import {
     DebugAdapterSessionFactoryImpl
 } from './debug-adapter';
 import { MessagingService } from '@theia/core/lib/node/messaging/messaging-service';
+import { ConnectionContainerModule } from '@theia/core/lib/node/messaging/messaging-contribution';
 import {
     DebugAdapterContribution,
     DebugAdapterSessionFactory,
     DebugAdapterFactory
 } from './debug-model';
+import { DebugServiceImpl } from './debug-service-impl';
+import { DebugAdapterContributionRegistry } from './debug-adapter-contribution-registry';
+import { DebugAdapterSessionManager } from './debug-adapter-session-manager';
 
-export default new ContainerModule(bind => {
-    bind(DebugService).to(DebugServiceImpl).inSingletonScope();
-    bind(MessagingService.Contribution).toService(DebugService);
-
-    bind(DebugAdapterSessionFactory).to(DebugAdapterSessionFactoryImpl).inSingletonScope();
-    bind(DebugAdapterFactory).to(LaunchBasedDebugAdapterFactory).inSingletonScope();
-    bind(DebugAdapterContributionRegistry).toSelf().inSingletonScope();
-    bind(DebugAdapterSessionManager).toSelf().inSingletonScope();
+const debugConnectionModule = new ContainerModule(bind => {
     bindContributionProvider(bind, DebugAdapterContribution);
+    bind(DebugAdapterContributionRegistry).toSelf().inSingletonScope();
 
+    bind(DebugService).to(DebugServiceImpl).inSingletonScope();
     bind(ConnectionHandler).toDynamicValue(context =>
         new JsonRpcConnectionHandler(DebugPath, () =>
             context.container.get<DebugService>(DebugService)
         )
     ).inSingletonScope();
+});
+
+export default new ContainerModule(bind => {
+    bind(ConnectionContainerModule).toConstantValue(debugConnectionModule);
+
+    bind(DebugAdapterSessionFactory).to(DebugAdapterSessionFactoryImpl).inSingletonScope();
+    bind(DebugAdapterFactory).to(LaunchBasedDebugAdapterFactory).inSingletonScope();
+    bind(DebugAdapterSessionManager).toSelf().inSingletonScope();
+    bind(MessagingService.Contribution).toService(DebugAdapterSessionManager);
 
     bind(ILogger).toDynamicValue(({ container }) =>
         container.get<ILogger>(ILogger).child('debug')
